requirements.txt
termcolor==1.1.0
ipaddress==1.0.22
ujson==1.35
pyzmq==17.1.2
redis==2.10.6
tornado==5.0.1
pycrypto==2.6.1
mysql-connector-python==8.0.12
GitPython==2.1.7
tormysql==0.3.7
Sphinx==1.6.3
pyOpenSSL==17.2.0
cffi==1.10.0
cryptography==2.3.1
futures==3.1.1
ipgetter==0.6
expiringdict==1.1.4
python-geoip==1.2
python-geoip-geolite2-yplan==2017.608
psutil==5.2.2
lazy==1.3
pympler==0.5
sprockets-influxdb==2.1.0
aioredis==1.1.0

git+https://github.com/anthill-utils/PyMySQL.git#egg=PyMySQL
git+https://github.com/anthill-utils/pyjwt.git#egg=PyJWT
git+https://github.com/anthill-utils/pika.git#egg=pika
